requirements.txt
Django>=3.2.3,<3.3
gunicorn==20.1.0
requests==2.28.2
Werkzeug==2.2.2
google-cloud-spanner==4.0.1